Camera を Sprite に変更
Camera を Sprite から継承するように
enchantjs に関連する処理を移動
