+++ samples/ja/make-rpg/hackforplay/camera.js
@@ -1,169 +1,40 @@
-import { Node } from 'enchantjs/enchant';
+import {
+	Core,
+	Node,
+	Event,
+	Sprite,
+	Surface,
+	Group
+} from 'enchantjs/enchant';
 import 'enchantjs/ui.enchant';
+import 'enchantjs/fix';
 import 'hackforplay/rpg-kit-main';
-
-import EventEmitter from 'mod/EventEmitter3';
 
 const clamp = function(value, min, max) {
 	return Math.max(min, Math.min(max, value));
 };
 
-enchant.Map.prototype.cvsRender = function cvsRender(ctx) {
 
-	if (this.width !== 0 && this.height !== 0) {
-		var core = enchant.Core.instance;
-		this.updateBuffer();
-		ctx.save();
-		ctx.setTransform(1, 0, 0, 1, 0, 0);
-		var cvs = this._context.canvas;
-		ctx.drawImage(cvs, 0, 0 /* , core.width, core.height */ );
-		ctx.restore();
-	}
-};
-
-enchant.Map.prototype.redraw = function(x, y, width, height) {
-
-	x = 0;
-	y = 0;
-	width = this.width;
-	height = this.height;
-
-	var core = enchant.Core.instance;
-	var surface = new enchant.Surface(width, height);
-	this._surface = surface;
-	var canvas = surface._element;
-	canvas.style.position = 'absolute';
-	if (enchant.ENV.RETINA_DISPLAY && core.scale === 2) {
-		canvas.width = width * 2;
-		canvas.height = height * 2;
-		this._style.webkitTransformOrigin = '0 0';
-		this._style.webkitTransform = 'scale(0.5)';
-	} else {
-		canvas.width = width;
-		canvas.height = height;
-	}
-	this._context = canvas.getContext('2d');
-
-
-	if (this._image == null) {
-		return;
-	}
-	var image, tileWidth, tileHeight, dx, dy;
-	if (this._doubledImage) {
-		image = this._doubledImage;
-		tileWidth = this._tileWidth * 2;
-		tileHeight = this._tileHeight * 2;
-		dx = -this._offsetX * 2;
-		dy = -this._offsetY * 2;
-		x *= 2;
-		y *= 2;
-		width *= 2;
-		height *= 2;
-	} else {
-		image = this._image;
-		tileWidth = this._tileWidth;
-		tileHeight = this._tileHeight;
-		dx = -this._offsetX;
-		dy = -this._offsetY;
-	}
-	var row = image.width / tileWidth | 0;
-	var col = image.height / tileHeight | 0;
-	var left = Math.max((x + dx) / tileWidth | 0, 0);
-	var top = Math.max((y + dy) / tileHeight | 0, 0);
-	var right = Math.ceil((x + dx + width) / tileWidth);
-	var bottom = Math.ceil((y + dy + height) / tileHeight);
-
-	var source = image._element;
-	var context = this._context;
-	var canvas = context.canvas;
-
-
-	context.clearRect(x, y, width, height);
-	for (var i = 0, len = this._data.length; i < len; i++) {
-		var data = this._data[i];
-		var r = Math.min(right, data[0].length);
-		var b = Math.min(bottom, data.length);
-		for (y = top; y < b; y++) {
-			for (x = left; x < r; x++) {
-				var n = data[y][x];
-				if (0 <= n && n < row * col) {
-					var sx = (n % row) * tileWidth;
-					var sy = (n / row | 0) * tileHeight;
-					context.drawImage(source, sx, sy, tileWidth, tileHeight,
-						x * tileWidth - dx, y * tileHeight - dy, tileWidth, tileHeight);
-				}
-			}
-		}
-	}
-
-}
-
-enchant.Event.RESIZE = 'resize';
-enchant.Event.RENDERED = 'rendered';
-
-
-enchant.CanvasRenderer.prototype.listener = new EventEmitter();
-enchant.CanvasRenderer.prototype.override = null;
-
-enchant.CanvasRenderer.instance.render = function(context, node, event) {
-
-	// safari 対策
-	if (!node.scene && !node._scene) return;
-
-	context = this.override || context;
-
-	// render start
-	this.listener.emit('renderStart', node);
-
-	enchant.CanvasRenderer.prototype.render.call(this, context, node, event);
-
-	// render end
-	this.listener.emit('renderEnd', node);
-
-	node.dispatchEvent(new enchant.Event(enchant.Event.RENDERED));
-};
-
-
-var init = window.RPGMap.prototype.initialize;
-window.RPGMap.prototype.initialize = function(ux, uy, x, y) {
-
-	init.apply(this, arguments);
-
-	// FIX PATCH
-	x = x !== undefined ? x : 15;
-	y = y !== undefined ? y : 10;
-	// ---- tera
-
-	this._surface = new enchant.Surface(ux * x, uy * y);
-
-};
-
-enchant.CanvasRenderer.instance.listener.on('renderStart', (node) => {
-
-
-
-	if (!Hack.map || node !== enchant.Core.instance.rootScene._layers.Canvas) return;
-
-	enchant.CanvasRenderer.instance.override = Hack.map._surface.context;
-
-});
-
-
-class Camera extends Node {
+class Camera extends Sprite {
 
 	constructor(x, y, w, h) {
-		super();
+		super(w, h);
 
+		// this.opacity = 0.5;
 
-		this.width = w || game.width;
-		this.height = h || game.height;
+		w = w || game.width;
+		h = h || game.height;
+
+		this.image = new Surface(w, h);
+
+		this.w = w;
+		this.h = h;
 
 		this.x = x || 0;
 		this.y = y || 0;
 
+
 		this.background = '#000';
-
-
 
 
 		this.enabled = true;
@@ -179,12 +50,25 @@ class Camera extends Node {
 		this.borderColor = '#000';
 		this.borderLineWidth = 1;
 
-
-
 		Camera.collection.push(this);
+
+		Hack.cameraGroup.addChild(this);
 
 	}
 
+
+	get w() { return this.width; }
+	get h() { return this.height; }
+
+	set w(value) {
+		this.width = value;
+		this.image._element.width = value;
+	}
+
+	set h(value) {
+		this.height = value;
+		this.image._element.height = value;
+	}
 
 
 	resize(width, height) {
@@ -213,30 +97,19 @@ class Camera extends Node {
 	}
 
 	getCenter() {
+
+		// center 固定
 		if (this.center) return this.center;
 
-		if (this.target) {
-
-			if (this.target instanceof RPGObject) {
-
-				return {
-
-					x: this.target.x - this.target.offset.x + 16,
-					y: this.target.y - this.target.offset.y + 16
-
-				};
-
-			}
-
-			return {
-				x: this.target.x,
-				y: this.target.y
-			};
+		// target
+		if (this.target && this.target instanceof RPGObject) {
+			return this.target.center;
 		}
 
+		// マップの中心
 		if (Hack.map) {
 
-			var map = Hack.map;
+			const map = Hack.map;
 
 			return {
 				x: map.width / 2,
@@ -245,31 +118,21 @@ class Camera extends Node {
 
 		}
 
-
-		return {
-			x: 0,
-			y: 0
-		};
-
+		console.error('Camera#getCenter');
 	}
-
 
 	getScale() {
 
-		var scale = this.scale;
+		// クリップしない
+		if (!this.clipScaleFunction) return this.scale;
 
-		if (this.clipScaleFunction) {
+		const x = Hack.map.width / this.w;
+		const y = Hack.map.height / this.h;
 
-			var x = Hack.map.width / this.width;
-			var y = Hack.map.height / this.height;
-			var clip = this.clipScaleFunction(x, y);
+		const clip = this.clipScaleFunction(x, y);
+		if (this.scale > clip) return clip;
 
-			if (scale > clip) scale = clip;
-
-		}
-
-
-		return scale;
+		return this.scale;
 	}
 
 
@@ -309,8 +172,7 @@ class Camera extends Node {
 	// 描画範囲を画面に収める
 	clampRect(rect) {
 
-		var w = this.getVisionSize().width;
-		var h = this.getVisionSize().height;
+		const { w, h } = this.getVisionSize();
 
 		var over = false;
 
@@ -385,9 +247,10 @@ class Camera extends Node {
 
 
 	getVisionSize() {
+		const scale = this.getScale();
 		return {
-			width: this.width * this.getScale(),
-			height: this.height * this.getScale()
+			w: this.w * scale,
+			h: this.h * scale
 		};
 	}
 
@@ -406,23 +269,14 @@ class Camera extends Node {
 
 	render() {
 
-		var ctx = enchant.Core.instance.rootScene._layers.Canvas.context;
 
 		var center = this.getCenter();
+
+		if (!center) return;
+
 		var x = center.x;
 		var y = center.y;
 
-
-		if (this.rootCanvasRendering) {
-			ctx.drawImage(Hack.map._surface._element, 0, 0);
-		}
-
-
-		/*
-		ctx.beginPath();
-		ctx.arc(x, y, 60, 0, Math.PI*2, false);
-		ctx.stroke();
-		*/
 
 		var rect = this.getRenderRect();
 		var r = rect;
@@ -430,15 +284,6 @@ class Camera extends Node {
 
 		// ctx.strokeRect(r.x, r.y, r.width, r.height);
 
-		ctx.save();
-
-		ctx.translate(this.x, this.y);
-
-		if (this.clip) {
-			ctx.beginPath();
-			ctx.rect(0, 0, this.width, this.height);
-			ctx.clip();
-		}
 
 		this.dispatchEvent((function() {
 
@@ -446,51 +291,34 @@ class Camera extends Node {
 
 			e.rect = r;
 
-			e.context = ctx;
 
 			return e;
 
 		})());
 
 
-		if (this.background) {
 
-			ctx.fillStyle = this.background;
-			ctx.fillRect(0, 0, this.width, this.height);
+		this.image.context.drawImage(
+			Hack.map._surface._element,
 
-		}
+			r.x, r.y, r.width, r.height,
+			0, 0, this.w, this.h);
 
+		/*
 
+		this.dispatchEvent(new Event('postrender', {
+			texture: this.image,
+			context: this.image.context,
+			rect: r,
+		}));
 
-		ctx.drawImage(Hack.map._surface._element, r.x, r.y, r.width, r.height, 0, 0, this.width, this.height);
-
-
-		if (this.border) {
-			ctx.strokeStyle = this.borderColor;
-			ctx.lineWidth = this.borderLineWidth;
-			ctx.strokeRect(0, 0, this.width, this.height);
-		}
-
-		this.dispatchEvent((function() {
-
-			var e = new enchant.Event(enchant.Event.RENDERED);
-
-			e.rect = r;
-
-			e.context = ctx;
-			return e;
-
-		})());
+		*/
 
 
+		this.image.context.fillStyle = 'red';
+		this.image.context.fillRect(0, 0, 100, 100);
 
-		if (this.clip) {
-			ctx.beginPath();
-			ctx.rect(0, 0, game.width, game.height);
-			ctx.clip();
-		}
 
-		ctx.restore();
 
 	}
 
@@ -501,22 +329,6 @@ Camera.collection = [];
 
 Camera.background = '#000';
 
-
-enchant.CanvasRenderer.instance.listener.on('renderEnd', function(node) {
-	if (!Hack.map || node !== Hack.map.scene) return;
-
-	enchant.CanvasRenderer.instance.override = null;
-	enchant.Core.instance.rootScene._layers.Canvas.context.fillStyle = Camera.background;
-	enchant.Core.instance.rootScene._layers.Canvas.context.fillRect(0, 0, game.width, game.height);
-
-
-	Camera.collection.forEach(function(camera) {
-
-		camera.render();
-
-	});
-
-});
 
 
 // カメラを並べる
@@ -559,17 +371,8 @@ Camera.arrange = function(x, y, border, filter) {
 
 Camera.layout = Camera.arrange;
 
-var camera = new Camera();
-camera.resize(game.width, game.height);
-Hack.camera = camera;
-
-game.on('load', function() {
-
-	// ターゲットが指定されていない場合はHack.playerになる
-	Hack.camera.target = Hack.camera.target || Hack.player;
-
-});
-
 
 window.Camera = Camera;
 Camera.main = Hack.camera;
+
+export default Camera;
